app_sdk/requirements.txt
aiohttp
pyyaml
asteval
pycrypto
